Minor refactor Documentation
