+++ kuberfest/commands/deploy.py
@@ -10,13 +10,14 @@ def run(project):
     Debug.info("Setting up kubernetes yamls...")
 
     yaml_tool = YamlTool(project)
+    kubernetes_tool = KubernetesTool(project)
 
     project.delete_tmp_dir()
     yamls = list()
     for yaml_file_name in settings.template_file_names:
         yamls.append(yaml_tool.get_yaml(yaml_file_name, **project.get_variables()))
     yaml_tool.merge_yamls(yamls, settings.output_yaml_file_name)
-    yaml_tool.run_yaml(settings.output_yaml_file_name)
+    kubernetes_tool.run_yaml(settings.output_yaml_file_name)
 
     # Wait for deployments
     deployments = list()
@@ -24,7 +25,6 @@ def run(project):
     for deployment_data in deployments_data:
         deployments.append(deployment_data['metadata']['name'])
 
-    kubernetes_tool = KubernetesTool(project)
     return kubernetes_tool.wait_for_deployments(
         namespace=project.get_variable('NAMESPACE'),
         deployments=deployments,
